+++ src/mmalloc/mmap-sup.c
@@ -316,8 +316,8 @@ __mmalloc_mmap_morecore(struct mdesc *mdp, int size)
   if (mdp->fd > 0)
     {
       /* FIXME:  Test results of lseek() and write() here: */
-      lseek(mdp->fd, (foffset + mapbytes - 1), SEEK_SET);
-      write(mdp->fd, &buf, 1);
+      lseek(mdp->fd, (off_t)(foffset + mapbytes - 1L), SEEK_SET);
+      write(mdp->fd, &buf, (size_t)1UL);
 
 #ifdef MAP_FILE
       flags = (MAP_PRIVATE_OR_SHARED(mdp) | MAP_FILE);
